requirements.txt
asgiref==3.5.2
boto3==1.23.2
botocore==1.26.2
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==4.0.4
django-cloudinary-storage==0.3.0
django-storages==1.12.3
djangorestframework==3.13.1
idna==3.3
jmespath==1.0.0
Pillow==9.1.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
